Remove bullet font customization & separate component spacing modifier
